requirements.txt
exceptiongroup==1.3.0
iniconfig==2.1.0
joblib==1.5.1
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
style==1.1.0
threadpoolctl==3.6.0
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
update==0.0.1
